[#667] Cleanup: remove deprecated endpoints and views
Streamlined code by eliminating unused account routes and view classes:
* Deleted obsolete functions (pricing, signup, login, signin, etc.)
* Removed views related to info, dashboard, and security policies
* Added a specific view for handling unknown secrets
* Improved maintainability by reducing redundant code paths
