+++ lib/onetime/app/web/views.rb
@@ -52,6 +52,15 @@ module Onetime
         self.template_extension = 'txt'
       end
 
+      class UnknownSecret < Onetime::App::View
+        self.template_name = :index
+        def init
+          self[:title] = "No such secret"
+          self[:display_feedback] = false
+          self[:display_masthead] = false
+        end
+      end
+
     end
   end
 end
